+++ src/widget/widget.rs
@@ -303,3 +303,4 @@ impl PushrodWidget for PushrodBaseWidget {
         );
     }
 }
+